tests/wpt/tests/accname/basic.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>

<!--

These tests should remain solid and passing in any implementation that supports get_computed_label.

It uses a standard promise_test (rather than aria-util.js) to reduce other dependencies.

If you're adding something you expect to fail in one or more implementations, you probably want a different file.

-->

<div id='d' style='height: 100px; width: 100px' role="group" aria-label="test label"></div>
<h1 id="h">test heading</h1>
<script>

promise_test(async t => {
  const label = await test_driver.get_computed_label(document.getElementById('d'));
  assert_equals(label, "test label");
}, "tests labelFrom: author");

promise_test(async t => {
  const label = await test_driver.get_computed_label(document.getElementById('h'));
  assert_equals(label, "test heading");
}, "tests labelFrom: contents");

</script>
